Peer: Refactor Peer.
Remove unnecessary Peer abstraction and move it to RoomManager
package. Also remove dummy package.
