--- app/src/main/java/com/nevilleantony/prototype/peer/PeerType.java
@@ -1,6 +1,0 @@
-package com.nevilleantony.prototype.peer;
-
-public enum PeerType {
-	DEFAULT,
-	ADMIN
-}
